Add trailing slash to kubelet location
This is one of those small things that are hard to notice

It took me a while to realise that there was something wrong with the path of kubelet. I use k0s and I had `/var/lib/k0s/kubelet` for the path of kubelet as stated in the docs. However, openebs doesn't add a missing trailing slash when prefixing to the full path, and so the registrar wasn't working properly